+++ src/Firebase/Auth/SignInWithIdpCredentials.php
@@ -11,6 +11,7 @@ final class SignInWithIdpCredentials implements IsTenantAware, SignIn
     private ?string $idToken = null;
     private ?string $linkingIdToken = null;
     private ?string $oauthTokenSecret = null;
+    private ?string $rawNonce = null;
     private string $requestUri = 'http://localhost';
     private ?TenantId $tenantId = null;
 
@@ -39,6 +40,14 @@ final class SignInWithIdpCredentials implements IsTenantAware, SignIn
     {
         $instance = new self($provider);
         $instance->idToken = $idToken;
+
+        return $instance;
+    }
+
+    public function withRawNonce(string $rawNonce): self
+    {
+        $instance = clone $this;
+        $instance->rawNonce = $rawNonce;
 
         return $instance;
     }
@@ -85,6 +94,11 @@ final class SignInWithIdpCredentials implements IsTenantAware, SignIn
     public function idToken(): ?string
     {
         return $this->idToken;
+    }
+
+    public function rawNonce(): ?string
+    {
+        return $this->rawNonce;
     }
 
     public function linkingIdToken(): ?string